Create Assignment 1 Warm-Up
In the full Assignment 1, you will implement an abstract 2-player board-game that we call Chomp++, inspired by the classical Chomp game. Your implementation will not (initially, anyway!) have to play the game; instead it will be maintaining the state of the game-board, checking that players’ moves are legal, and then displaying the result of the move. 

This warm-up component is a relatively simple exercise that motivates you to write some code, which will be part of your implementation for the full Assignment 1.

Chomp++ Description / Rules
Chomp++ is played on a 2D grid. Initially the grid is a rectangle, but this rectangle gets progressively smaller as the game proceeds, and may not remain a rectangle.   A player’s move is to “point” at a cell on the board. Every call above and to the right of that cell and the pointed cell are then chomped. This removes all those cells from the board.  For example, if the initial state is a 4x5 rectangle, it looks like:

         0    1    2    3    4 

   0  |    |    |    |    |    |
   1  |    |    |    |    |    |
   2  |    |    |    |    |    |
   3  |    |    |    |    |    |
where I have added row and column numbers .

The first player's move might be to point at cell (2,3) (using 
(
r
o
w
,
c
o
l
u
m
n
)
(row,column))).  The result would be:

         0    1    2    3    4 

   0  |    |    |    |XXXXXXXXXX
   1  |    |    |    |XXXXXXXXXX
   2  |    |    |    |XXXXXXXXXX
   3  |    |    |    |    |    |
where I've used the X characters to indicate cells that have been chomped.  If the second player then chomps at (3,4), the result would be 

         0    1    2    3    4

   0  |    |    |    |XXXXXXXXXX
   1  |    |    |    |XXXXXXXXXX
   2  |    |    |    |XXXXXXXXXX
   3  |    |    |    |    |XXXXX
The game finishes when the board has been completely destroyed/chomped.

Scoring and Cell Contents
So far, so boring.  The interest comes from the fact that the board is not entirely empty. Cells can contain coins (good), poison (bad), and antidote (stops poison). When you chomp a bit of the board, you get the contents of all the chomped cells.  If you chomp coins, they accrue and will eventually be your final score, if you don't die.  If you chomp poison, you die (the game ends early and the other player wins).  But: if you already have enough antidote to counteract all the poison you chomped, the antidote is used up, and you survive after all.

The Warmup Task: Part 1
You will implement functions that will be used to adjust a player’s state based on the list of what they've chomped.  Any one cell is either Empty, or contains some quantity of coins, poison or antidote.  Thus the Haskell data type is

data CellContent = Coins Int | Poison Int | Antidote Int | Empty
If you have just chomped a number of cells, you get a list of CellContent values (one element in the list for each cell).  Given, such a list, you need to calculate

totalCoins: the total number of coins (this will be added to the player’s accumulating stash);

totalPoison: the total poison count (this will need to be compared with the player’s current antidote level); and 

totalAntidote: the total antidote count (this will be added to the player’s current antidote level)


The Warmup Task: Part 2
The final part of the warmup is to write the function that adjusts a player state by applying the effect of a list of chomped cells.  The player state is another simple user data type:

data PlayerState = PS_ac Int Int
The _ac suffix is to remind you that the first number is the antidote count, and the second number is the coin count.  (It’s in alphabetical order if that helps.)

The final part of the task is to write the adjustPlayerState function. This takes the starting state and a list of CellContent values (the “chomped list”) and returns the updated player state.  The first (life and death) rule is:

the player dies (return Nothing) if the total quantity of poison chomped exceeds the starting antidote count;

We are using the Maybe type to indicate whether or not the player lives.  We don’t care how many coins or antidote doses a player has if they die, so a dead player can just be recorded as Nothing.  If the player lives, the return value will be Just newvalue, with newvalue being calculated according to the following rules:

the player’s coin count is increased by the number of coins in the chomped list;

the player’s antidote count is decreased by the number of poisons in the chomped list, and then increased by the number of antidotes in the chomped list

We give two examples.

If a player's state is PS_ac 1 3 initially, and then they chomp [Antidote 2, Poison 1, Poison 2], the player will die, because the newly chomped poison count exceeds the initial antidote count.

If a player's state is PS_ac 4 10 initially, and then they chomp [Coins 4, Poison 2, Empty, Antidote 6, Poison 1, Coins 5], the player survives. Three of the initial antidotes are consumed to counteract the newly chomped poisons. The player's new state is PS_ac 7 19.